Remove IMG_{Get,Set}Error wrappers
In the original C headers, IMG_GetError() is simply a macro that expands
to SDL_GetError(). So far, in these Pascal units, IMG_GetError() was a
Pascal function that called SDL_GetError() internally. This commit
removes the Pascal function and instead defines IMG_GetError as an
alternative name linking to SDL_GetError().

The same is done for IMG_SetError().