Remove Mix_{Clear,Get,Set}Error wrappers
In the original C headers, Mix_GetError() is simply a macro that expands
to SDL_GetError(). So far, in these Pascal units, Mix_GetError() was a
Pascal function that called SDL_GetError() internally. This commit
removes the Pascal function and instead defines Mix_GetError as an
alternative name linking to SDL_GetError().

The same is done for Mix_SetError() and Mix_ClearError().