Remove function wrappers in SDL2_ttf.pas
In the original C headers, the following functions do not really exist;
they are just macros that expand to the name of a different function:
- TTF_GetError() -> SDL_GetError()
- TTF_SetError() -> SDL_SetError()
- TTF_RenderText() -> TTF_RenderText_Shaded()
- TTF_RenderUTF8() -> TTF_RenderUTF8_Shaded()
- TTF_RenderUNICODE() -> TTF_RenderUNICODE_Shaded()
So far, in these Pascal units, all of the above were recreated as Pascal
functions that called the appropriate "real" function. This commit
removes the Pascal functions and instead defines each of the above
as an alternative name linking to the appropriate function.